Fixed DS3231_intalarm.ino example
The use of "alarmState" variable was not needed. The status of the LED can be changed with the help of "isAlarm" variable.
Also used in setup function the "clock.enableOutput(false)" to set the INTCN register to 1 ( condition for triggering the alarm interrupt ).
Made the "attachInterrupt" function call more explicit.